CalculateAverageResultOfYearProject
This C++ console app calculates the average result for a year project. You input the number of materials, each material’s name, the number of exams, teacher’s points, and attention percentage (e.g., 25%), along with the material’s coefficient. The app then generates a table with exam points, material averages, and the overall average. Note that grades range from 1 to 20, with a maximum of 4 exams per material. Data is saved in AverageResultOfYearData.txt, which acts as the database.